+++ home/screens/LocationDiagnosticsScreen.tsx
@@ -6,7 +6,15 @@ import * as Permissions from 'expo-permissions';
 import * as TaskManager from 'expo-task-manager';
 import { EventEmitter, EventSubscription } from 'fbemitter';
 import * as React from 'react';
-import { AppState, AppStateStatus, Platform, StyleSheet, Text, View } from 'react-native';
+import {
+  AppState,
+  AppStateStatus,
+  NativeEventSubscription,
+  Platform,
+  StyleSheet,
+  Text,
+  View,
+} from 'react-native';
 import MapView, { Polyline } from 'react-native-maps';
 
 import NavigationEvents from '../components/NavigationEvents';
@@ -46,6 +54,7 @@ export default class LocationDiagnosticsScreen extends React.Component<Props, St
   mapViewRef = React.createRef<MapView>();
 
   eventSubscription?: EventSubscription;
+  appStateSubscription?: NativeEventSubscription;
 
   readonly state: State = {
     isBackgroundLocationAvailable: null,
@@ -70,7 +79,7 @@ export default class LocationDiagnosticsScreen extends React.Component<Props, St
     const { status } = await Permissions.askAsync(Permissions.LOCATION);
 
     if (status !== 'granted') {
-      AppState.addEventListener('change', this.handleAppStateChange);
+      this.appStateSubscription = AppState.addEventListener('change', this.handleAppStateChange);
       this.setState({
         error:
           'Location permissions are required in order to use this feature. You can manually enable them at any time in the "Location Services" section of the Settings app.',
@@ -114,7 +123,10 @@ export default class LocationDiagnosticsScreen extends React.Component<Props, St
     }
 
     if (this.state.initialRegion) {
-      AppState.removeEventListener('change', this.handleAppStateChange);
+      if (this.appStateSubscription) {
+        this.appStateSubscription.remove();
+        this.appStateSubscription = undefined;
+      }
       return;
     }
 
@@ -126,7 +138,10 @@ export default class LocationDiagnosticsScreen extends React.Component<Props, St
       this.eventSubscription.remove();
     }
 
-    AppState.removeEventListener('change', this.handleAppStateChange);
+    if (this.appStateSubscription != null) {
+      this.appStateSubscription.remove();
+      this.appStateSubscription = undefined;
+    }
   }
 
   async startLocationUpdates(accuracy = this.state.accuracy) {